pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "latamgpt-benchmark"
version = "0.1.0"
description = "Batch evaluation pipeline for CHOCLO and Trueque with OpenAI, Doubleword, and W&B Weave."
readme = "README.md"
requires-python = ">=3.12"
license = { file = "LICENSE" }
authors = [
  { name = "Alonso" },
]
dependencies = [
  "datasets>=4.3.0",
  "openai>=2.6.1",
  "python-dotenv>=1.1.1",
  "wandb>=0.23.0",
  "weave>=0.52.9",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4.2",
  "ruff>=0.14.2",
]

[project.scripts]
latamgpt-benchmark = "latamgpt_benchmark.cli:main"
latamgpt-judge-only = "latamgpt_benchmark.judge_only:main"

[tool.hatch.build.targets.wheel]
packages = ["src/latamgpt_benchmark"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 100
